+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Some rolls using math functions could cause errors
 - Item activation data was hidden
+- Context notes weren't showing up on item sheets
 
 ## 0.78.0